pyproject.toml
[project]
name = "VitAI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.4",
    "azure-ai-inference",
    "gradio>=4.0.0",
]
